// applications/v5/frontend/src/app/shared/models/api.models.ts
export interface ErrorResponse {
  timestamp: string;
  status: number;
  error: string;
  code?: string;
  message: string;
  path: string;
  details?: string[];
}

export interface FieldError {
  field: string;
  message: string;
}

export interface ValidationErrorResponse extends ErrorResponse {
  fieldErrors: FieldError[];
}